# app/api/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Fabric CLI http wrapper"
  },
  "host": "localhost",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "definitions": {
    "Channels": {
      "type": "object",
      "properties": {
        "status": {
          "default": 200,
          "type": "number"
        },
        "channels": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "channel_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "curChannel": {
      "type": "object",
      "properties": {
        "currentChannel": {
          "type": "string"
        }
      }
    },
    "blockPayload": {
      "type": "object",
      "properties": {
        "header": {
          "type": "object",
          "properties": {
            "channel_header": {
              "type": "object",
              "properties": {
                "version": {
                  "type": "number"
                },
                "type": {
                  "type": "number"
                },
                "time_stamp": {
                  "type": "string"
                },
                "channel_id": {
                  "type": "string"
                },
                "txid": {
                  "type": "string"
                },
                "epoch": {
                  "type": "string"
                }
              }
            },
            "signature_header": {
              "type": "object",
              "properties": {
                "creator": {
                  "type": "object",
                  "properties": {
                    "Mspid": {
                      "type": "string"
                    },
                    "Idbytes": {
                      "type": "string"
                    }
                  }
                },
                "nonce": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "number"
                      }
                    },
                    "type": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "data": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "paths": {
...
    "/orgs/docker": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "operationId": "getDockerAtrifacts",
        "summary": "Get docker artifact for new org",
        "description": "Get docker artifact for new org",
        "parameters": [
          {
            "name": "orderer",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "newOrg",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "numPeers",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "randomNumber",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "produces": ["application/zip"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/orgs/switch": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "operationId": "switchOrg",
        "summary": "Switch to different org",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/orgs/addToChannel": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "operationId": "addOrgToChannel",
        "summary": "Add new org to channel",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/orgs/addToConsortium": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "operationId": "addToConsortium",
        "summary": "Add new org to consortium",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/chaincode": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "operationId": "installChaincode",
        "summary": "Chaincode install",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/channel/{channel}/chaincode": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "operationId": "instantiateChaincode",
        "summary": "instantiate chaincode",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "channel",
            "in": "path",
            "description": "channel",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/channels": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "channelList",
        "summary": "Query to fetch channels",
        "description": "Query to fetch channels",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Channels"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/curChannel": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "currentChannel",
        "summary": "get Current channel",
        "description": "get Current channel",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/curChannel"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/changeChannel/{channel_genesis_hash}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "getChannelChange",
        "summary": "Query to get Change Channel on a channel",
        "description": "Query to get Change Channel on a channel",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "currentChannel": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/chaincode/{channel_genesis_hash}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Chaincodes"],
        "operationId": "chaincodeList",
        "summary": "Query to fetch all Installed/instantiated chaincodes",
        "description": "Query to fetch all Installed/instantiated chaincodes",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "path": {
                    "type": "string"
                  },
                  "channel_genesis_hash": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/channel": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "tags": ["Channel"],
        "operationId": "createChannel",
        "summary": "Create New channel on target peers  ",
        "description": "Create New channel on target peers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "channelName",
            "in": "formData",
            "description": "channel name",
            "required": true,
            "type": "string"
          },
          {
            "name": "genesisBlock",
            "in": "formData",
            "description": "gensis block",
            "required": true,
            "type": "string"
          },
          {
            "name": "orgName",
            "in": "formData",
            "description": "organation name ",
            "required": true,
            "type": "string"
          },
          {
            "name": "channelArtificats",
            "in": "formData",
            "description": "channelArtificats",
            "required": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "name": "profile",
            "in": "formData",
            "description": "block profile",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/joinChannel": {
      "x-swagger-router-controller": "platformroutes",
      "post": {
        "tags": ["Channel"],
        "operationId": "joinChannel",
        "summary": "join channel ",
        "description": "join  channel ",
        "consumes": ["application/json", "application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "payload",
            "schema": {
              "type": "object",
              "properties": {
                "channelName": {
                  "type": "string"
                },
                "orgName": {
                  "type": "string"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/block/{channel_genesis_hash}/{number}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Blocks"],
        "operationId": "blockByNumber",
        "summary": "Query to Fetch Blocks",
        "description": "Query to Fetch Blocks",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          },
          {
            "name": "number",
            "in": "path",
            "description": "block number",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "number": {
                  "type": "string"
                },
                "previous_hash": {
                  "type": "string"
                },
                "channel_genesis_hash": {
                  "type": "string"
                },
                "data_hash": {
                  "type": "string"
                },
                "transactions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "signature": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string"
                          },
                          "data": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          }
                        }
                      },
                      "payload": {
                        "$ref": "#/definitions/blockPayload"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/peers/{channel_genesis_hash}": {
      "x-swagger-router-controller": "dbroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "peerList",
        "summary": "Peers",
        "description": "Peers",
        "parameters": [
          {
            "name": "channel_genesis_hash",
            "in": "path",
            "description": "channel_genesis_hash",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "requests": {
                        "type": "string"
                      },
                      "channel_genesis_hash": {
                        "type": "string"
                      },
                      "server_hostname": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/peersStatus/{channelName}": {
      "x-swagger-router-controller": "platformroutes",
      "get": {
        "tags": ["Channel"],
        "operationId": "peerList",
        "summary": "Query Get PeersStatus",
        "description": "Query Get PeersStatus",
        "parameters": [
          {
            "name": "channelName",
            "in": "path",
            "description": "channel name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number"
                },
                "peers": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "requests": {
                        "type": "string"
                      },
                      "server_hostname": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  }
}